+++ src/assets/src/configurator/GroupSettings.js
@@ -71,13 +71,6 @@ export default Settings.extend({
     return this._id
   },
 
-  /**
-   * @deprecated in 3.8.0
-   */
-  setSortOrder (_) {
-    console.warn('GroupSettings.setSortOrder() is deprecated and no longer used.')
-  },
-
   getName () { return this._name ?? this._originalSettings.name },
   setName (name) {
     if (name !== this._name) {